use Long for all ids(fixes #11)

+++ app/src/main/java/com/vobbla16/mesh/data/remote/MeshApi.kt
@@ -61,7 +61,7 @@ class MeshApi : KoinComponent {
         }
 
     suspend fun getHomework(
-        studentId: Int,
+        studentId: Long,
         beginDate: String,
         endDate: String
     ) =
